refactor: move created_at formatting from model to widget
`TextNote::created_at` returned a locale/timezone-dependent `HH:MM:SS` string,
mixing a presentation concern into the model layer and pulling a `chrono`
dependency into the model. It also made the value untestable without depending
on the machine's timezone.

Have the model expose the raw `Timestamp` and move the local `%T` wall-clock
formatting into `TextNoteWidget` as a `formatted_created_at` method, next to the
other view logic. The model no longer depends on `chrono`.
